requirements.txt
faiss-cpu~=1.11.0
streamlit~=1.46.1
openai~=1.93.0
python-dotenv~=1.1.1
sentence-transformers~=4.1.0
numpy~=2.2.6
pandas~=2.3.0